--- app/api/utils.py
@@ -1,50 +1,0 @@
-import smtplib
-import jwt
-from jwt.exceptions import InvalidTokenError
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
-from app.core.config import settings
-from app.core import security
-from datetime import datetime, timedelta, timezone
-
-
-def send_email(*, email_to: str, subject: str, content: str):
-    msg = MIMEMultipart()
-    msg["From"] = settings.SMTP_USER
-    msg["To"] = email_to
-    msg["Subject"] = subject
-
-    msg.attach(MIMEText(content, "plain"))
-
-    try:
-        server = smtplib.SMTP(settings.SMTP_HOST, settings.SMTP_PORT)
-        server.starttls()
-        server.login(settings.SMTP_USER, settings.SMTP_PASSWORD)
-        server.sendmail(settings.SMTP_USER, email_to, msg.as_string())
-        server.quit()
-        print("Email sent successfully!")
-    except Exception as e:
-        print(f"Error sending email: {e}")
-
-
-def verify_password_reset_token(token: str) -> str | None:
-    try:
-        decoded_token = jwt.decode(
-            token, settings.SECRET_KEY, algorithms=[security.ALGORITHM]
-        )
-        return str(decoded_token["sub"])
-    except InvalidTokenError:
-        return None
-
-
-def generate_password_reset_token(email: str) -> str:
-    delta = timedelta(hours=settings.EMAIL_RESET_TOKEN_EXPIRE_HOURS)
-    now = datetime.now(timezone.utc)
-    expires = now + delta
-    exp = expires.timestamp()
-    encoded_jwt = jwt.encode(
-        {"exp": exp, "nbf": now, "sub": email},
-        settings.SECRET_KEY,
-        algorithm=security.ALGORITHM,
-    )
-    return encoded_jwt
